Changed the rest of the dropToken tests to use expected toString
changed testDropToken2InAColumn.

changed testDropTokenFillRow
added a commenet to testDropTokenFillRow as well to make changes to documentation and changes to code to make to account for documentation changes.

changed. testDropTokenFillColumn.

added a commenet to testDropTokenFillColumn as well to make changes to documentation and changes to code to make to account for documentation changes.

changed testDropTokenFillBoard