fixed testIsPlayerAtPosSingleToken, added contract for GameBoardMem isPlayerAtPos
removed assertFalse statement.
the second assert statement is redundant. if the first assertTrue passed then the assertFalse with the wrong player token would pass you are effectively testing the same thing twice. Also the original test broke your code when testing a player that is not a key in hashmap thus a contract was added to account for that.